requirements.txt
pystray~=0.19.5
pillow~=11.2.0
PyGObject~=3.52.3